+++ js/utils.js
@@ -38,6 +38,7 @@ const getRandomItemsArray = (arr, count) => {
 };
 
 const ALERT_SHOW_TIME = 5000;
+// Функция, описывающая стиль всплывающего сообщения
 const showAlert = (message) => {
   const alertContainer = document.createElement('div');
   alertContainer.style.zIndex = 100;
